Add blog controller and REST routes
